requirements.txt
astroid==2.2.5
autopep8==1.4.3
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
entrypoints==0.3
ffmpeg-python==0.1.17
flake8==3.7.7
flatten-json==0.1.6
future==0.17.1
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-speech==1.0.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.3
pandas==0.24.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pydub==0.23.1
pyflakes==2.1.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rope==0.14.0
rsa==4.0
simpleaudio==1.0.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
youtube-dl==2019.7.16
